refactor: improved code base for easier understanding
