+++ internal/client/constants.go
@@ -1,0 +1,3 @@
+package infisicalclient
+
+const USER_AGENT = "terraform"
