+++ internal/crypto/model.go
@@ -1,0 +1,7 @@
+package crypto
+
+type SymmetricEncryptionResult struct {
+	CipherText []byte `json:"CipherText"`
+	Nonce      []byte `json:"Nonce"`
+	AuthTag    []byte `json:"AuthTag"`
+}
