+++ internal/provider/resource/project_user_resource.go
@@ -1,9 +1,9 @@
-package provider
+package resource 
 
 import (
 	"context"
 	"fmt"
-	infisical "terraform-provider-infisical/client"
+	infisical "terraform-provider-infisical/internal/client"
 	"time"
 
 	"github.com/hashicorp/terraform-plugin-framework/path"
@@ -17,35 +17,38 @@ import (
 
 // Ensure the implementation satisfies the expected interfaces.
 var (
-	_ resource.Resource = &ProjectIdentityResource{}
+	_ resource.Resource = &ProjectUserResource{}
 )
 
 // NewProjectResource is a helper function to simplify the provider implementation.
-func NewProjectIdentityResource() resource.Resource {
-	return &ProjectIdentityResource{}
+func NewProjectUserResource() resource.Resource {
+	return &ProjectUserResource{}
 }
 
-// ProjectIdentityResource is the resource implementation.
-type ProjectIdentityResource struct {
+// ProjectUserResource is the resource implementation.
+type ProjectUserResource struct {
 	client *infisical.Client
 }
 
 // projectResourceSourceModel describes the data source data model.
-type ProjectIdentityResourceModel struct {
-	ProjectID    types.String          `tfsdk:"project_id"`
-	IdentityID   types.String          `tfsdk:"identity_id"`
-	Identity     types.Object          `tfsdk:"identity"`
-	Roles        []ProjectIdentityRole `tfsdk:"roles"`
-	MembershipId types.String          `tfsdk:"membership_id"`
+type ProjectUserResourceModel struct {
+	ProjectID    types.String      `tfsdk:"project_id"`
+	Username     types.String      `tfsdk:"username"`
+	User         types.Object      `tfsdk:"user"`
+	Roles        []ProjectUserRole `tfsdk:"roles"`
+	MembershipId types.String      `tfsdk:"membership_id"`
 }
 
-type ProjectIdentityDetails struct {
-	ID         types.String `tfsdk:"id"`
-	Name       types.String `tfsdk:"name"`
-	AuthMethod types.String `tfsdk:"auth_method"`
+type ProjectUserPersonalDetails struct {
+	ID        types.String `tfsdk:"id"`
+	Email     types.String `tfsdk:"email"`
+	FirstName types.String `tfsdk:"first_name"`
+	LastName  types.String `tfsdk:"last_name"`
 }
 
-type ProjectIdentityRole struct {
+const TEMPORARY_MODE_RELATIVE = "relative"
+
+type ProjectUserRole struct {
 	ID                      types.String `tfsdk:"id"`
 	RoleSlug                types.String `tfsdk:"role_slug"`
 	CustomRoleID            types.String `tfsdk:"custom_role_id"`
@@ -57,12 +60,12 @@ type ProjectIdentityRole struct {
 }
 
 // Metadata returns the resource type name.
-func (r *ProjectIdentityResource) Metadata(_ context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_project_identity"
+func (r *ProjectUserResource) Metadata(_ context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_project_user"
 }
 
 // Schema defines the schema for the resource.
-func (r *ProjectIdentityResource) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
+func (r *ProjectUserResource) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		Description: "Create projects & save to Infisical. Only Machine Identity authentication is supported for this data source",
 		Attributes: map[string]schema.Attribute{
@@ -70,41 +73,44 @@ func (r *ProjectIdentityResource) Schema(_ context.Context, _ resource.SchemaReq
 				Description: "The id of the project",
 				Required:    true,
 			},
-			"identity_id": schema.StringAttribute{
-				Description: "The id of the identity.",
+			"username": schema.StringAttribute{
+				Description: "The usename of the user. By default its the email",
 				Required:    true,
 			},
 			"membership_id": schema.StringAttribute{
-				Description:   "The membershipId of the project identity",
+				Description:   "The membershipId of the project user",
 				Computed:      true,
 				PlanModifiers: []planmodifier.String{stringplanmodifier.UseStateForUnknown()},
 			},
-			"identity": schema.SingleNestedAttribute{
+			"user": schema.SingleNestedAttribute{
 				Computed:    true,
-				Description: "The identity details of the project identity",
+				Description: "The user details of the project user",
 				Attributes: map[string]schema.Attribute{
-					"id": schema.StringAttribute{
-						Description:   "The ID of the identity",
-						Computed:      true,
-						PlanModifiers: []planmodifier.String{stringplanmodifier.UseStateForUnknown()},
-					},
-					"name": schema.StringAttribute{
-						Description: "The name of the identity",
+					"email": schema.StringAttribute{
+						Description: "The email of the user",
 						Computed:    true,
 					},
-					"auth_method": schema.StringAttribute{
-						Description: "The auth method for the identity",
+					"first_name": schema.StringAttribute{
+						Description: "The first name of the user",
+						Computed:    true,
+					},
+					"last_name": schema.StringAttribute{
+						Description: "The last name of the user",
+						Computed:    true,
+					},
+					"id": schema.StringAttribute{
+						Description: "The id of the user",
 						Computed:    true,
 					},
 				},
 			},
 			"roles": schema.ListNestedAttribute{
 				Required:    true,
-				Description: "The roles assigned to the project identity",
+				Description: "The roles assigned to the project user",
 				NestedObject: schema.NestedAttributeObject{
 					Attributes: map[string]schema.Attribute{
 						"id": schema.StringAttribute{
-							Description: "The ID of the project identity role.",
+							Description: "The ID of the project user role.",
 							Computed:    true,
 						},
 						"role_slug": schema.StringAttribute{
@@ -150,7 +156,7 @@ func (r *ProjectIdentityResource) Schema(_ context.Context, _ resource.SchemaReq
 }
 
 // Configure adds the provider configured client to the resource.
-func (r *ProjectIdentityResource) Configure(_ context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
+func (r *ProjectUserResource) Configure(_ context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
 	if req.ProviderData == nil {
 		return
 	}
@@ -170,24 +176,24 @@ func (r *ProjectIdentityResource) Configure(_ context.Context, req resource.Conf
 }
 
 // Create creates the resource and sets the initial Terraform state.
-func (r *ProjectIdentityResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+func (r *ProjectUserResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 	if r.client.Config.AuthStrategy != infisical.AuthStrategy.UNIVERSAL_MACHINE_IDENTITY {
 		resp.Diagnostics.AddError(
-			"Unable to create project identity",
+			"Unable to create project user",
 			"Only Machine Identity authentication is supported for this operation",
 		)
 		return
 	}
 
 	// Retrieve values from plan
-	var plan ProjectIdentityResourceModel
+	var plan ProjectUserResourceModel
 	diags := req.Plan.Get(ctx, &plan)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	var roles []infisical.CreateProjectIdentityRequestRoles
+	var roles []infisical.UpdateProjectUserRequestRoles
 	var hasAtleastOnePermanentRole bool
 	for _, el := range plan.Roles {
 		isTemporary := el.IsTemporary.ValueBool()
@@ -219,7 +225,7 @@ func (r *ProjectIdentityResource) Create(ctx context.Context, req resource.Creat
 			temporaryRange = "1h"
 		}
 
-		roles = append(roles, infisical.CreateProjectIdentityRequestRoles{
+		roles = append(roles, infisical.UpdateProjectUserRequestRoles{
 			Role:                     el.RoleSlug.ValueString(),
 			IsTemporary:              isTemporary,
 			TemporaryMode:            temporaryMode,
@@ -228,38 +234,50 @@ func (r *ProjectIdentityResource) Create(ctx context.Context, req resource.Creat
 		})
 	}
 	if !hasAtleastOnePermanentRole {
-		resp.Diagnostics.AddError("Error assigning role to identity", "Must have atleast one permanent role")
+		resp.Diagnostics.AddError("Error assigning role to user", "Must have atleast one permanent role")
 		return
 	}
 
-	_, err := r.client.CallCreateProjectIdentity(infisical.CreateProjectIdentityRequest{
-		ProjectID:  plan.ProjectID.ValueString(),
-		IdentityID: plan.IdentityID.ValueString(),
-		Roles:      roles,
+	_, err := r.client.InviteUsersToProject(infisical.InviteUsersToProjectRequest{
+		ProjectID: plan.ProjectID.ValueString(),
+		Usernames: []string{plan.Username.ValueString()},
 	})
 	if err != nil {
 		resp.Diagnostics.AddError(
-			"Error attaching identity to project",
+			"Error inviting user",
 			"Couldn't save project to Infiscial, unexpected error: "+err.Error(),
 		)
 		return
 	}
 
-	projectIdentityDetails, err := r.client.CallGetProjectIdentityByID(infisical.GetProjectIdentityByIDRequest{
-		ProjectID:  plan.ProjectID.ValueString(),
-		IdentityID: plan.IdentityID.ValueString(),
+	projectUserDetails, err := r.client.GetProjectUserByUsername(infisical.GetProjectUserByUserNameRequest{
+		ProjectID: plan.ProjectID.ValueString(),
+		Username:  plan.Username.ValueString(),
 	})
 	if err != nil {
 		resp.Diagnostics.AddError(
-			"Error fetching identity",
-			"Couldn't find identity in project, unexpected error: "+err.Error(),
+			"Error fetching user",
+			"Couldn't find user in project, unexpected error: "+err.Error(),
 		)
 		return
 	}
 
-	planRoles := make([]ProjectIdentityRole, 0, len(projectIdentityDetails.Membership.Roles))
-	for _, el := range projectIdentityDetails.Membership.Roles {
-		val := ProjectIdentityRole{
+	updatedRoles, err := r.client.UpdateProjectUser(infisical.UpdateProjectUserRequest{
+		ProjectID:    plan.ProjectID.ValueString(),
+		MembershipID: projectUserDetails.Membership.ID,
+		Roles:        roles,
+	})
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Error assigning roles to user",
+			"Couldn't update role , unexpected error: "+err.Error(),
+		)
+		return
+	}
+
+	planRoles := make([]ProjectUserRole, 0, len(updatedRoles.Roles))
+	for _, el := range updatedRoles.Roles {
+		val := ProjectUserRole{
 			ID:                      types.StringValue(el.ID),
 			RoleSlug:                types.StringValue(el.Role),
 			TemporaryAccessEndTime:  types.StringValue(el.TemporaryAccessEndTime.Format(time.RFC3339)),
@@ -278,19 +296,20 @@ func (r *ProjectIdentityResource) Create(ctx context.Context, req resource.Creat
 		planRoles = append(planRoles, val)
 	}
 	plan.Roles = planRoles
-	plan.MembershipId = types.StringValue(projectIdentityDetails.Membership.ID)
+	plan.MembershipId = types.StringValue(projectUserDetails.Membership.ID)
 	diags = resp.State.Set(ctx, plan)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	identityDetails := ProjectIdentityDetails{
-		ID:         types.StringValue(projectIdentityDetails.Membership.Identity.Id),
-		Name:       types.StringValue(projectIdentityDetails.Membership.Identity.Name),
-		AuthMethod: types.StringValue(projectIdentityDetails.Membership.Identity.AuthMethod),
+	userPersonalDetails := ProjectUserPersonalDetails{
+		Email:     types.StringValue(projectUserDetails.Membership.User.Email),
+		FirstName: types.StringValue(projectUserDetails.Membership.User.FirstName),
+		LastName:  types.StringValue(projectUserDetails.Membership.User.LastName),
+		ID:        types.StringValue(projectUserDetails.Membership.User.ID),
 	}
-	diags = resp.State.SetAttribute(ctx, path.Root("identity"), identityDetails)
+	diags = resp.State.SetAttribute(ctx, path.Root("user"), userPersonalDetails)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
@@ -299,37 +318,38 @@ func (r *ProjectIdentityResource) Create(ctx context.Context, req resource.Creat
 }
 
 // Read refreshes the Terraform state with the latest data.
-func (r *ProjectIdentityResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+func (r *ProjectUserResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
 	if r.client.Config.AuthStrategy != infisical.AuthStrategy.UNIVERSAL_MACHINE_IDENTITY {
 		resp.Diagnostics.AddError(
-			"Unable to read project identity",
+			"Unable to read project user",
 			"Only Machine Identity authentication is supported for this operation",
 		)
 		return
 	}
 
 	// Get current state
-	var state ProjectIdentityResourceModel
+	var state ProjectUserResourceModel
 	diags := req.State.Get(ctx, &state)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	projectIdentityDetails, err := r.client.CallGetProjectIdentityByID(infisical.GetProjectIdentityByIDRequest{
-		ProjectID:  state.ProjectID.ValueString(),
-		IdentityID: state.IdentityID.ValueString(),
+	projectUserDetails, err := r.client.GetProjectUserByUsername(infisical.GetProjectUserByUserNameRequest{
+		ProjectID: state.ProjectID.ValueString(),
+		Username:  state.Username.ValueString(),
 	})
 	if err != nil {
 		resp.Diagnostics.AddError(
-			"Error fetching identity",
-			"Couldn't find identity in project, unexpected error: "+err.Error(),
+			"Error fetching user",
+			"Couldn't find user in project, unexpected error: "+err.Error(),
 		)
 		return
 	}
-	planRoles := make([]ProjectIdentityRole, 0, len(projectIdentityDetails.Membership.Roles))
-	for _, el := range projectIdentityDetails.Membership.Roles {
-		val := ProjectIdentityRole{
+
+	planRoles := make([]ProjectUserRole, 0, len(projectUserDetails.Membership.Roles))
+	for _, el := range projectUserDetails.Membership.Roles {
+		val := ProjectUserRole{
 			ID:                      types.StringValue(el.ID),
 			RoleSlug:                types.StringValue(el.Role),
 			TemporaryAccessEndTime:  types.StringValue(el.TemporaryAccessEndTime.Format(time.RFC3339)),
@@ -355,12 +375,13 @@ func (r *ProjectIdentityResource) Read(ctx context.Context, req resource.ReadReq
 		return
 	}
 
-	identityDetails := ProjectIdentityDetails{
-		ID:         types.StringValue(projectIdentityDetails.Membership.Identity.Id),
-		Name:       types.StringValue(projectIdentityDetails.Membership.Identity.Name),
-		AuthMethod: types.StringValue(projectIdentityDetails.Membership.Identity.AuthMethod),
+	userPersonalDetails := ProjectUserPersonalDetails{
+		Email:     types.StringValue(projectUserDetails.Membership.User.Email),
+		FirstName: types.StringValue(projectUserDetails.Membership.User.FirstName),
+		LastName:  types.StringValue(projectUserDetails.Membership.User.LastName),
+		ID:        types.StringValue(projectUserDetails.Membership.User.ID),
 	}
-	diags = resp.State.SetAttribute(ctx, path.Root("identity"), identityDetails)
+	diags = resp.State.SetAttribute(ctx, path.Root("user"), userPersonalDetails)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
@@ -369,39 +390,39 @@ func (r *ProjectIdentityResource) Read(ctx context.Context, req resource.ReadReq
 }
 
 // Update updates the resource and sets the updated Terraform state on success.
-func (r *ProjectIdentityResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+func (r *ProjectUserResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
 	if r.client.Config.AuthStrategy != infisical.AuthStrategy.UNIVERSAL_MACHINE_IDENTITY {
 		resp.Diagnostics.AddError(
-			"Unable to update project identity",
+			"Unable to update project user",
 			"Only Machine Identity authentication is supported for this operation",
 		)
 		return
 	}
 
 	// Retrieve values from plan
-	var plan ProjectIdentityResourceModel
+	var plan ProjectUserResourceModel
 	diags := req.Plan.Get(ctx, &plan)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	var state ProjectIdentityResourceModel
+	var state ProjectUserResourceModel
 	diags = req.State.Get(ctx, &state)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	if state.IdentityID != plan.IdentityID {
+	if state.Username != plan.Username {
 		resp.Diagnostics.AddError(
-			"Unable to update project identity",
-			fmt.Sprintf("Cannot change identity id, previous identity: %s, new identity id: %s", state.IdentityID, plan.IdentityID),
+			"Unable to update project user",
+			fmt.Sprintf("Cannot change username, previous username: %s, new username: %s", state.Username, plan.Username),
 		)
 		return
 	}
 
-	var roles []infisical.UpdateProjectIdentityRequestRoles
+	var roles []infisical.UpdateProjectUserRequestRoles
 	var hasAtleastOnePermanentRole bool
 	for _, el := range plan.Roles {
 		isTemporary := el.IsTemporary.ValueBool()
@@ -433,7 +454,7 @@ func (r *ProjectIdentityResource) Update(ctx context.Context, req resource.Updat
 			temporaryRange = "1h"
 		}
 
-		roles = append(roles, infisical.UpdateProjectIdentityRequestRoles{
+		roles = append(roles, infisical.UpdateProjectUserRequestRoles{
 			Role:                     el.RoleSlug.ValueString(),
 			IsTemporary:              isTemporary,
 			TemporaryMode:            temporaryMode,
@@ -443,38 +464,38 @@ func (r *ProjectIdentityResource) Update(ctx context.Context, req resource.Updat
 	}
 
 	if !hasAtleastOnePermanentRole {
-		resp.Diagnostics.AddError("Error assigning role to identity", "Must have atleast one permanent role")
+		resp.Diagnostics.AddError("Error assigning role to user", "Must have atleast one permanent role")
 		return
 	}
 
-	projectIdentityDetails, err := r.client.CallGetProjectIdentityByID(infisical.GetProjectIdentityByIDRequest{
-		ProjectID:  plan.ProjectID.ValueString(),
-		IdentityID: plan.IdentityID.ValueString(),
+	projectUserDetails, err := r.client.GetProjectUserByUsername(infisical.GetProjectUserByUserNameRequest{
+		ProjectID: plan.ProjectID.ValueString(),
+		Username:  plan.Username.ValueString(),
 	})
 	if err != nil {
 		resp.Diagnostics.AddError(
-			"Error fetching identity",
-			"Couldn't find identity in project, unexpected error: "+err.Error(),
+			"Error fetching user",
+			"Couldn't find user in project, unexpected error: "+err.Error(),
 		)
 		return
 	}
 
-	updatedRoles, err := r.client.CallUpdateProjectIdentity(infisical.UpdateProjectIdentityRequest{
-		ProjectID:  plan.ProjectID.ValueString(),
-		IdentityID: projectIdentityDetails.Membership.Identity.Id,
-		Roles:      roles,
+	updatedRoles, err := r.client.UpdateProjectUser(infisical.UpdateProjectUserRequest{
+		ProjectID:    plan.ProjectID.ValueString(),
+		MembershipID: projectUserDetails.Membership.ID,
+		Roles:        roles,
 	})
 	if err != nil {
 		resp.Diagnostics.AddError(
-			"Error assigning roles to identity",
+			"Error assigning roles to user",
 			"Couldn't update role , unexpected error: "+err.Error(),
 		)
 		return
 	}
 
-	planRoles := make([]ProjectIdentityRole, 0, len(updatedRoles.Roles))
+	planRoles := make([]ProjectUserRole, 0, len(updatedRoles.Roles))
 	for _, el := range updatedRoles.Roles {
-		val := ProjectIdentityRole{
+		val := ProjectUserRole{
 			ID:                      types.StringValue(el.ID),
 			RoleSlug:                types.StringValue(el.Role),
 			TemporaryAccessEndTime:  types.StringValue(el.TemporaryAccessEndTime.Format(time.RFC3339)),
@@ -499,12 +520,13 @@ func (r *ProjectIdentityResource) Update(ctx context.Context, req resource.Updat
 		return
 	}
 
-	identityDetails := ProjectIdentityDetails{
-		ID:         types.StringValue(projectIdentityDetails.Membership.Identity.Id),
-		Name:       types.StringValue(projectIdentityDetails.Membership.Identity.Name),
-		AuthMethod: types.StringValue(projectIdentityDetails.Membership.Identity.AuthMethod),
+	userPersonalDetails := ProjectUserPersonalDetails{
+		Email:     types.StringValue(projectUserDetails.Membership.User.Email),
+		FirstName: types.StringValue(projectUserDetails.Membership.User.FirstName),
+		LastName:  types.StringValue(projectUserDetails.Membership.User.LastName),
+		ID:        types.StringValue(projectUserDetails.Membership.User.ID),
 	}
-	diags = resp.State.SetAttribute(ctx, path.Root("identity"), identityDetails)
+	diags = resp.State.SetAttribute(ctx, path.Root("user"), userPersonalDetails)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
@@ -512,32 +534,32 @@ func (r *ProjectIdentityResource) Update(ctx context.Context, req resource.Updat
 }
 
 // Delete deletes the resource and removes the Terraform state on success.
-func (r *ProjectIdentityResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+func (r *ProjectUserResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
 
 	if r.client.Config.AuthStrategy != infisical.AuthStrategy.UNIVERSAL_MACHINE_IDENTITY {
 		resp.Diagnostics.AddError(
-			"Unable to delete project identity",
+			"Unable to delete project user",
 			"Only Machine Identity authentication is supported for this operation",
 		)
 		return
 	}
 
-	var state ProjectIdentityResourceModel
+	var state ProjectUserResourceModel
 	diags := req.State.Get(ctx, &state)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	_, err := r.client.CallDeleteProjectIdentity(infisical.DeleteProjectIdentityRequest{
-		ProjectID:  state.ProjectID.ValueString(),
-		IdentityID: state.IdentityID.ValueString(),
+	_, err := r.client.DeleteProjectUser(infisical.DeleteProjectUserRequest{
+		ProjectID: state.ProjectID.ValueString(),
+		Username:  []string{state.Username.ValueString()},
 	})
 
 	if err != nil {
 		resp.Diagnostics.AddError(
-			"Error deleting project identity",
-			"Couldn't delete project identity from Infiscial, unexpected error: "+err.Error(),
+			"Error deleting project user",
+			"Couldn't delete project user from Infiscial, unexpected error: "+err.Error(),
 		)
 		return
 	}
